Attempt to fix github issue #7
Issue #7 was filed for a program that didn't produce output
when only given one input line.  It is trying to append to
the pattern space with 'N', which has nothing to do if there
was only one line.

Thus, it appears that when the 'N' command runs out of input, it should
not signal EOF to the engine, so that the sed script can continue.
Making this change allows the example program from issue #7 to match my
system `sed`, and also does not break any of the test programs in
engine_test.go.

I added a test to engine_test.go with the example from issue #7
to make sure this fix never regresses.  And, since I was in there
I added another test of the branches and 't' command.